add fuzz search operator
