+++ packages/core/package.json
@@ -39,6 +39,7 @@
     "mathjs": "^12.4.2",
     "mime-types": "^2.1.35",
     "minimatch": "^9.0.4",
+    "minisearch": "^6.3.0",
     "openai": "^4.40.0",
     "prettier": "^3.2.5",
     "pretty-bytes": "^6.1.1",
