+++ packages/core/src/index.ts
@@ -34,6 +34,7 @@ export * from "./import"
 export * from "./fs"
 export * from "./tokens"
 export * from "./promptdom"
+export * from "./websearch"
 export * from "./search"
 export * from "./yaml"
 export * from "./mime"
