+++ packages/sample/genaisrc/style/genaiscript.d.ts
@@ -1016,7 +1016,7 @@ interface HighlightOptions {
     maxLength?: number
 }
 
-interface SearchResult {
+interface WebSearchResult {
     webPages: WorkspaceFile[]
 }
 
@@ -1041,12 +1041,44 @@ interface VectorSearchEmbeddingsOptions extends VectorSearchOptions {
     chunkOverlap?: number
 }
 
+interface FuzzSearchOptions {
+    /**
+     * Controls whether to perform fuzzy search. It can be a simple boolean, or a
+     * number, or a function.
+     *
+     * If a boolean is given, fuzzy search with a default fuzziness parameter is
+     * performed if true.
+     *
+     * If a number higher or equal to 1 is given, fuzzy search is performed, with
+     * a maximum edit distance (Levenshtein) equal to the number.
+     *
+     * If a number between 0 and 1 is given, fuzzy search is performed within a
+     * maximum edit distance corresponding to that fraction of the term length,
+     * approximated to the nearest integer. For example, 0.2 would mean an edit
+     * distance of 20% of the term length, so 1 character in a 5-characters term.
+     * The calculated fuzziness value is limited by the `maxFuzzy` option, to
+     * prevent slowdown for very long queries.
+     *
+     * If a function is passed, the function is called upon search with a search
+     * term, a positional index of that term in the tokenized search query, and
+     * the tokenized search query. It should return a boolean or a number, with
+     * the meaning documented above.
+     */
+    fuzzy?: boolean | number
+    /**
+     * Controls the maximum fuzziness when using a fractional fuzzy value. This is
+     * set to 6 by default. Very high edit distances usually don't produce
+     * meaningful results, but can excessively impact search performance.
+     */
+    maxFuzzy?: number
+}
+
 interface Retrieval {
     /**
      * Executers a Bing web search. Requires to configure the BING_SEARCH_API_KEY secret.
      * @param query
      */
-    webSearch(query: string): Promise<SearchResult>
+    webSearch(query: string): Promise<WebSearchResult>
 
     /**
      * Search using similiraty distance on embeddings
@@ -1073,6 +1105,18 @@ interface Retrieval {
      * @deprecated Use `vectorSearch` instead
      */
     search?: undefined
+
+    /**
+     * Performs a fuzzy search over the files
+     * @param query keywords to search
+     * @param files list of files
+     * @param options fuzzing configuration
+     */
+    fuzzSearch(
+        query: string,
+        files: (string | WorkspaceFile) | (string | WorkspaceFile)[],
+        options?: FuzzSearchOptions
+    ): Promise<WorkspaceFile[]>
 }
 
 type FetchTextOptions = Omit<RequestInit, "body" | "signal" | "window">
